files

+++ C++/codeforces/1734C.cpp
@@ -5,10 +5,10 @@ int main(){
     std::cin >> t;
     int resp[t] = {};
     for(int i = 0; i < t; i++){
-        int n = 0;
+        long n = 0;
         std::cin >> n;
         int binary[n] = {};
-        for(int j = 0; j < n; j++){
+        for(long j = 0; j < n; j++){
             int b = 0;
             std::cin >> b;
             binary[j] = b;
